Convert forceLightningInterfaceCreate
